+++ test/utils/Checkpoints.test.js
@@ -18,12 +18,19 @@ contract('Checkpoints', function (accounts) {
     describe('without checkpoints', function () {
       it('returns zero as latest value', async function () {
         expect(await this.checkpoint.latest()).to.be.bignumber.equal('0');
+
+        const ckpt = await this.checkpoint.latestCheckpoint();
+        expect(ckpt[0]).to.be.equal(false);
+        expect(ckpt[1]).to.be.bignumber.equal('0');
+        expect(ckpt[2]).to.be.bignumber.equal('0');
       });
 
       it('returns zero as past value', async function () {
         await time.advanceBlock();
-        expect(await this.checkpoint.getAtBlock(await web3.eth.getBlockNumber() - 1)).to.be.bignumber.equal('0');
-        expect(await this.checkpoint.getAtRecentBlock(await web3.eth.getBlockNumber() - 1)).to.be.bignumber.equal('0');
+        expect(await this.checkpoint.getAtBlock(await web3.eth.getBlockNumber() - 1))
+          .to.be.bignumber.equal('0');
+        expect(await this.checkpoint.getAtProbablyRecentBlock(await web3.eth.getBlockNumber() - 1))
+          .to.be.bignumber.equal('0');
       });
     });
 
@@ -39,9 +46,14 @@ contract('Checkpoints', function (accounts) {
 
       it('returns latest value', async function () {
         expect(await this.checkpoint.latest()).to.be.bignumber.equal('3');
+
+        const ckpt = await this.checkpoint.latestCheckpoint();
+        expect(ckpt[0]).to.be.equal(true);
+        expect(ckpt[1]).to.be.bignumber.equal(web3.utils.toBN(this.tx3.receipt.blockNumber));
+        expect(ckpt[2]).to.be.bignumber.equal(web3.utils.toBN('3'));
       });
 
-      for (const fn of [ 'getAtBlock(uint256)', 'getAtRecentBlock(uint256)' ]) {
+      for (const fn of [ 'getAtBlock(uint256)', 'getAtProbablyRecentBlock(uint256)' ]) {
         describe(`lookup: ${fn}`, function () {
           it('returns past values', async function () {
             expect(await this.checkpoint.methods[fn](this.tx1.receipt.blockNumber - 1)).to.be.bignumber.equal('0');
@@ -78,6 +90,18 @@ contract('Checkpoints', function (accounts) {
         expect(await this.checkpoint.length()).to.be.bignumber.equal(lengthBefore.addn(1));
         expect(await this.checkpoint.latest()).to.be.bignumber.equal('10');
       });
+
+      it('more than 5 checkpoints', async function () {
+        for (let i = 4; i <= 6; i++) {
+          await this.checkpoint.push(i);
+        }
+        expect(await this.checkpoint.length()).to.be.bignumber.equal('6');
+        const block = await web3.eth.getBlockNumber();
+        // recent
+        expect(await this.checkpoint.getAtProbablyRecentBlock(block - 1)).to.be.bignumber.equal('5');
+        // non-recent
+        expect(await this.checkpoint.getAtProbablyRecentBlock(block - 9)).to.be.bignumber.equal('0');
+      });
     });
   });
 
@@ -90,23 +114,27 @@ contract('Checkpoints', function (accounts) {
       describe('without checkpoints', function () {
         it('returns zero as latest value', async function () {
           expect(await this.contract.latest()).to.be.bignumber.equal('0');
+
+          const ckpt = await this.contract.latestCheckpoint();
+          expect(ckpt[0]).to.be.equal(false);
+          expect(ckpt[1]).to.be.bignumber.equal('0');
+          expect(ckpt[2]).to.be.bignumber.equal('0');
         });
 
         it('lookup returns 0', async function () {
           expect(await this.contract.lowerLookup(0)).to.be.bignumber.equal('0');
           expect(await this.contract.upperLookup(0)).to.be.bignumber.equal('0');
-          expect(await this.contract.upperLookupRecent(0)).to.be.bignumber.equal('0');
         });
       });
 
       describe('with checkpoints', function () {
         beforeEach('pushing checkpoints', async function () {
           this.checkpoints = [
-            { key: 2, value: '17' },
-            { key: 3, value: '42' },
-            { key: 5, value: '101' },
-            { key: 7, value: '23' },
-            { key: 11, value: '99' },
+            { key: '2', value: '17' },
+            { key: '3', value: '42' },
+            { key: '5', value: '101' },
+            { key: '7', value: '23' },
+            { key: '11', value: '99' },
           ];
           for (const { key, value } of this.checkpoints) {
             await this.contract.push(key, value);
@@ -114,8 +142,12 @@ contract('Checkpoints', function (accounts) {
         });
 
         it('returns latest value', async function () {
-          expect(await this.contract.latest())
-            .to.be.bignumber.equal(last(this.checkpoints).value);
+          expect(await this.contract.latest()).to.be.bignumber.equal(last(this.checkpoints).value);
+
+          const ckpt = await this.contract.latestCheckpoint();
+          expect(ckpt[0]).to.be.equal(true);
+          expect(ckpt[1]).to.be.bignumber.equal(last(this.checkpoints).key);
+          expect(ckpt[2]).to.be.bignumber.equal(last(this.checkpoints).value);
         });
 
         it('cannot push values in the past', async function () {
@@ -149,7 +181,6 @@ contract('Checkpoints', function (accounts) {
             const value = last(this.checkpoints.filter(x => i >= x.key))?.value || '0';
 
             expect(await this.contract.upperLookup(i)).to.be.bignumber.equal(value);
-            expect(await this.contract.upperLookupRecent(i)).to.be.bignumber.equal(value);
           }
         });
       });